@@ -76,10 +76,10 @@
     "rollup-plugin-gzip": "^2.2.0",
     "rollup-plugin-terser": "^7.0.0",
     "rollup-plugin-typescript2": "^0.27.2",
-    "rxjs": "^6.5.5",
+    "rxjs": "^7.2.0",
     "sinon": "^7.5.0",
     "ts-node": "^8.5.4",
-    "typescript": "^4.1.2",
+    "typescript": "^4.3.5",
     "version-bump-prompt": "^5.0.6"
   }
 }
